depreciated old way of extending vuex
