+++ core/app.ts
@@ -21,6 +21,7 @@ import VueApollo from 'vue-apollo'
 
 // TODO simplify by removing global mixins, plugins and filters - it can be done in normal 'vue' way
 import { registerTheme } from '@vue-storefront/core/lib/themes'
+import { registerModules } from '@vue-storefront/core/lib/module'
 import { prepareStoreView } from '@vue-storefront/store/lib/multistore'
 import { plugins, mixins, filters } from '@vue-storefront/core/compatibility/lib/themes'
 import { once } from '@vue-storefront/core/helpers'
@@ -30,15 +31,12 @@ import { Logger } from '@vue-storefront/core/lib/logger'
 // Entrys
 import App from 'theme/App.vue'
 import store from '@vue-storefront/store'
-// Will be depreciated in 1.7
-import themeModules from 'theme/store'
 
 import { enabledModules } from './modules-entry'
 
 // Will be depreciated in 1.8
 import { registerExtensions } from '@vue-storefront/core/compatibility/lib/extensions'
 import { registerExtensions as extensions } from 'src/extensions'
-
 
 function createRouter (): VueRouter {
   return new VueRouter({
@@ -57,16 +55,6 @@ function createRouter (): VueRouter {
       }
     }
   })
-}
-
-function registerModules (modules: VueStorefrontModule[], store: Store<RootState>, router: VueRouter, context): void {
-  let registeredModules = []
-  modules.forEach(m => registeredModules.push(m.register(store, router)))
-  Logger.info('VS Modules registration finished.', 'module', {
-      succesfulyRegistered: registeredModules.length + ' / ' + modules.length,
-      registrationOrder: registeredModules
-    }
-  )()
 }
 
 let router: VueRouter = null
@@ -90,15 +78,6 @@ function createApp (ssrContext, config): { app: Vue, router: VueRouter, store: S
   store.state.__DEMO_MODE__ = (config.demomode === true) ? true : false
   if(ssrContext) Vue.prototype.$ssrRequestContext = ssrContext
   if (!store.state.config) store.state.config = buildTimeConfig // if provided from SSR, don't replace it
-
-  // depreciated, will be removed in 1.7
-  const storeModules = themeModules || {}
-
-  // depreciated, will be removed in 1.7
-  for (const moduleName of Object.keys(storeModules)) {
-    console.debug('Registering Vuex module', moduleName)
-    store.registerModule(moduleName, storeModules[moduleName])
-  }
 
   const storeView = prepareStoreView(null) // prepare the default storeView
   store.state.storeView = storeView
