--- src/themes/default/resource/ig_feed.json
@@ -1,26 +1,0 @@
-[
-  {
-    "image": "/assets/ig/ig01.jpg",
-    "alt": "Woman practicing on the beach"
-  },
-  {
-    "image": "/assets/ig/ig02.jpg",
-    "alt": "Man practicing on the beach"
-  },
-  {
-    "image": "/assets/ig/ig03.jpg",
-    "alt": "Woman drinks water form the bottle on the beach"
-  },
-  {
-    "image": "/assets/ig/ig04.jpg",
-    "alt": "Man rests on the beach"
-  },
-  {
-    "image": "/assets/ig/ig05.jpg",
-    "alt": "Woman practicing on the beach"
-  },
-  {
-    "image": "/assets/ig/ig06.jpg",
-    "alt": "Man is holding a watter bottle"
-  }
-]
