--- src/themes/default/store/index.js
@@ -1,12 +1,0 @@
-// Here you can place theme-specific stores that will be added to the core stores.
-// Everything that you will export here will be treated as a Vuex module
-// The good practice is to keep all modulase separated in this folder.
-
-// Below you can find commented exaple for additional theme store
-
-// DEPRECIATED, will be removed in 1.7, use module.extend() instead
-import social from './social-tiles'
-
-export default {
-  social
-}
